Consolidate PGC specs into actions and observations modules.
Collapse base, simple_observations, wall_stats, and day_area_trace into actions.py and observations.py, and route action-trace updates through ActionSpec.trace_action.
