+++ tests/environments/PlantGrowthChamber/test_specs.py
@@ -1,8 +1,12 @@
+from unittest.mock import MagicMock
+
 import numpy as np
 
+from environments.PlantGrowthChamber.factory import ComposedPlantGrowthChamber
 from environments.PlantGrowthChamber.specs import (
     ACTION_SPECS,
     ColorTriangleAction,
+    DayAreaTraceObservation,
     DiscreteAction,
     IntensityAction,
     OneHotTimeObservation,
@@ -35,6 +39,53 @@ def test_color_triangle_trace_dim():
     assert create_observation_spec(
         "day_area_trace", ACTION_SPECS["color_triangle"], {}
     ).shape == (5,)
+
+
+def test_intensity_trace_action_decodes_scalar():
+    spec = IntensityAction()
+    result = spec.trace_action(0.8, backend=None)
+    np.testing.assert_allclose(result, BALANCED_ACTION_105 * 0.8)
+
+
+def test_color_triangle_trace_action_keeps_coefficients():
+    spec = ColorTriangleAction()
+    action = np.array([0.2, 0.5, 0.3])
+    np.testing.assert_array_equal(spec.trace_action(action, None), action)
+
+
+def test_color_triangle_trace_action_projects_six_channel():
+    spec = ColorTriangleAction()
+    result = spec.trace_action(np.zeros(6), None)
+    assert result.shape == (3,)
+
+
+def test_update_action_trace_decodes_before_uema_update():
+    backend = MagicMock()
+    backend.get_local_time.return_value.date.return_value = __import__(
+        "datetime"
+    ).date(2025, 1, 1)
+    obs_spec = DayAreaTraceObservation(trace_dim=6)
+    action_spec = IntensityAction()
+    env = ComposedPlantGrowthChamber(backend, obs_spec, action_spec, {})
+
+    env.update_action_trace(0.8)
+    assert obs_spec.action_uema is not None
+    trace = np.asarray(obs_spec.action_uema.compute()).reshape(-1)
+    np.testing.assert_allclose(trace, BALANCED_ACTION_105 * 0.8)
+
+
+def test_update_action_trace_handles_six_channel_night_action_for_color_triangle():
+    backend = MagicMock()
+    backend.get_local_time.return_value.date.return_value = __import__(
+        "datetime"
+    ).date(2025, 1, 1)
+    obs_spec = DayAreaTraceObservation(trace_dim=3)
+    action_spec = ColorTriangleAction()
+    env = ComposedPlantGrowthChamber(backend, obs_spec, action_spec, {})
+
+    env.update_action_trace(np.zeros(6))
+    assert obs_spec.action_uema is not None
+    assert np.asarray(obs_spec.action_uema.compute()).reshape(-1).shape == (3,)
 
 
 def test_one_hot_time_observation_shape():
